pyproject.toml
[project]
name = "ruche"
version = "0.1.0"
description = "Production-grade cognitive engine for conversational AI"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Ruche Team" }
]
keywords = ["ai", "cognitive", "conversational", "llm", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "prometheus-client>=0.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=1.20",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "numpy>=2.3.5",
    "scipy>=1.16.3",
    "scikit-learn>=1.7.2",
    "python-Levenshtein>=0.27.0", # Required for entity deduplication
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "python-jose>=3.3",
    "sse-starlette>=2.0",
    "httpx>=0.28",
    "redis>=7.1.0",
    "anthropic>=0.67.0",
    "openai>=2.8.1",
    "tiktoken>=0.12.0",
    "asyncpg>=0.31.0",
    "agno>=2.3.4",
    "qdrant-client>=1.16.1",
    "alembic>=1.17.2",
    "hatchet-sdk>=1.21.3",
    "rank-bm25>=0.2.2",
    "simpleeval>=1.0.3",
    "nest-asyncio>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "pydoc-markdown>=4.8.0",
]
# Optional providers - only install if you need them
embeddings = [
    "sentence-transformers>=5.1.2",  # For local embedding generation
]
queue = [
    "rq>=2.6.1",  # For Redis-based task queue
    "redis>=6.0.0",
]
api = [
    "redis>=6.0.0",  # For rate limiting and idempotency cache
]
# PostgreSQL production stores
postgres = [
    "asyncpg",  # PostgreSQL async driver
    "pgvector",  # PostgreSQL vector similarity extension
]
# Database migrations
migrations = [
    "alembic",  # Database schema migrations
    "asyncpg",  # Required for async Alembic
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-python-jose>=3.5.0.20250531",
    "jupyter>=1.1.1",
    "ipykernel>=7.1.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ruche"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "integration: marks tests as integration tests (require external services)",
    "performance: marks tests as performance tests",
    "slow: marks tests as slow running",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG001",  # Unused function argument (common in fixtures)
    "ARG002",  # Unused method argument (common in mocks)
    "B017",    # Assert blind exception (acceptable in tests)
]

[tool.ruff.lint.isort]
known-first-party = ["ruche"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
